[AIP-21] Move contrib.sensors.python_sensor sensors.python
