+++ airflow/contrib/sensors/python_sensor.py
@@ -16,58 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from typing import Callable, Dict, List, Optional
+"""This module is deprecated. Please use `airflow.sensors.python`."""
 
-from airflow.operators.python_operator import PythonOperator
-from airflow.sensors.base_sensor_operator import BaseSensorOperator
-from airflow.utils.decorators import apply_defaults
+import warnings
 
+# pylint: disable=unused-import
+from airflow.sensors.python import PythonSensor  # noqa
 
-class PythonSensor(BaseSensorOperator):
-    """
-    Waits for a Python callable to return True.
-
-    User could put input argument in templates_dict
-    e.g ``templates_dict = {'start_ds': 1970}``
-    and access the argument by calling ``kwargs['templates_dict']['start_ds']``
-    in the callable
-
-    :param python_callable: A reference to an object that is callable
-    :type python_callable: python callable
-    :param op_kwargs: a dictionary of keyword arguments that will get unpacked
-        in your function
-    :type op_kwargs: dict
-    :param op_args: a list of positional arguments that will get unpacked when
-        calling your callable
-    :type op_args: list
-    :param templates_dict: a dictionary where the values are templates that
-        will get templated by the Airflow engine sometime between
-        ``__init__`` and ``execute`` takes place and are made available
-        in your callable's context after the template has been applied.
-    :type templates_dict: dict of str
-    """
-
-    template_fields = ('templates_dict', 'op_args', 'op_kwargs')
-
-    @apply_defaults
-    def __init__(
-            self,
-            python_callable: Callable,
-            op_args: Optional[List] = None,
-            op_kwargs: Optional[Dict] = None,
-            templates_dict: Optional[Dict] = None,
-            *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.python_callable = python_callable
-        self.op_args = op_args or []
-        self.op_kwargs = op_kwargs or {}
-        self.templates_dict = templates_dict
-
-    def poke(self, context: Dict):
-        context.update(self.op_kwargs)
-        context['templates_dict'] = self.templates_dict
-        self.op_kwargs = PythonOperator.determine_op_kwargs(self.python_callable, context, len(self.op_args))
-
-        self.log.info("Poking callable: %s", str(self.python_callable))
-        return_value = self.python_callable(*self.op_args, **self.op_kwargs)
-        return bool(return_value)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.sensors.python`.",
+    DeprecationWarning, stacklevel=2
+)
