+++ scripts/ci/pylint_todo.txt
@@ -86,7 +86,7 @@
 ./airflow/providers/apache/hdfs/sensors/hdfs.py
 ./airflow/providers/jira/sensors/jira.py
 ./airflow/providers/mongo/sensors/mongo.py
-./airflow/contrib/sensors/python_sensor.py
+./airflow/sensors/python.py
 ./airflow/contrib/sensors/qubole_sensor.py
 ./airflow/contrib/sensors/redis_key_sensor.py
 ./airflow/contrib/sensors/redis_pub_sub_sensor.py
